% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boundary.Service.R
\name{Boundary.Service}
\alias{Boundary.Service}
\title{Boundary Service}
\value{
A \code{list} of services:
\itemize{
\item \code{upper()}: Calculates the upper boundary for outlier detection.
\item \code{lower()}: Calculates the lower boundary for outlier detection.
}
}
\description{
\code{Boundary.Service()} is a function that returns a set of services used to calculate the upper and lower boundaries for outlier detection.
It uses the inter quartile range (IQR) method and the skewness of the data to determine the boundaries.
}
