% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boundary.Validator.R
\name{Boundary.Validator}
\alias{Boundary.Validator}
\title{Boundary Validator}
\value{
A \code{list} of validators:
\itemize{
\item \code{sample()}
\item \code{exist()}
\item \code{is.numeric()}
}
}
\description{
\code{Boundary.Validator()} is an internal function that returns a set of validators used by \code{Boundary.Service()}.
\itemize{
\item The first validator: \code{sample}, checks if the sample data provided is not NULL and is numeric.
\item The second validator: \code{exist}, checks if a non-null data argument was provided.
\item The third validator: \code{is.numeric}, checks if the data argument provided contains numerical data.
}
}
