% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\name{box_plot_univariate}
\alias{box_plot_univariate}
\title{Return a vector with outlier scores based on the number of outlier feature and the median distance between extreme value}
\usage{
box_plot_univariate(data, func = distance)
}
\arguments{
\item{data}{a data frame containing only numerical features}

\item{func}{function of distance (univariate)}
}
\description{
Return a vector with outlier scores based on the number of outlier feature and the median distance between extreme value
}
