% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cat2bool}
\alias{cat2bool}
\title{A function to convert categorical variables into booleans}
\usage{
cat2bool(data)
}
\arguments{
\item{data}{the original dataset}
}
\description{
returns a dataframe with the categorical data replaced by booleans
}
