% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrmax.R
\name{corrmax}
\alias{corrmax}
\title{corrmax}
\usage{
corrmax(population.data, sample.data)
}
\arguments{
\item{population.data}{should be a target data set, e.g., the original raw data but only for the samples within the robust structure}

\item{sample.data}{should be a reference data set, e.g., the original data.}
}
