% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\name{distance}
\alias{distance}
\title{Compute the relative distance between x and extreme value, 0 if x is in the interval}
\usage{
distance(x, min_value, max_value)
}
\arguments{
\item{x}{value of the point to check}

\item{min_value}{minimum value of the interval}

\item{max_value}{maximum value of the interval}
}
\description{
Compute the relative distance between x and extreme value, 0 if x is in the interval
}
