% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{ensemble_function}
\alias{ensemble_function}
\title{Compute the score of each points based on the score it made it every functions}
\usage{
ensemble_function(lor, fun = mean, nrot = 50, val, dist_funct = TRUE)
}
\arguments{
\item{lor}{list of list of results named "probability_vector"}

\item{v_char}{vector containing the name of the functions}

\item{draw}{boolean to print and draw}
}
\description{
Compute the score of each points based on the score it made it every functions
}
