% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outlier.Exposure.R
\name{extract.outliers}
\alias{extract.outliers}
\title{Extract outliers}
\usage{
extract.outliers(input, name)
}
\arguments{
\item{input}{A vector or data.frame from which to extract outliers.}

\item{name}{The name of the column in the data.frame which should be used for outlier extraction. If \code{input} is a vector, this parameter is ignored.}
}
\value{
A vector or data.frame with outliers.
}
\description{
\code{Outliers.extract()} is a function that extracts outliers from a sample.
}
