% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\name{extreme_studentized_deviate}
\alias{extreme_studentized_deviate}
\title{Extreme studentized deviate univariate method. 
code adapted from # from http://www.itl.nist.gov/div898/handbook/eda/section3/eda35h3.r}
\usage{
extreme_studentized_deviate(x, alpha = 0.05)
}
\arguments{
\item{x}{a variable}

\item{alpha}{alpha}
}
\description{
Extreme studentized deviate univariate method. 
code adapted from # from http://www.itl.nist.gov/div898/handbook/eda/section3/eda35h3.r
}
