% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{feature_bagging}
\alias{feature_bagging}
\title{do feature_bagging}
\usage{
feature_bagging(data, bagging = "mi", factor = 0.5)
}
\arguments{
\item{bagging}{the bagging method}

\item{d}{the original dataset}
}
\description{
do feature_bagging
}
