% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\name{func_feature}
\alias{func_feature}
\title{Return a boolean matrix: evaluate the function func for each points of data}
\usage{
func_feature(data, func = function(x, min_value, max_value) {     (x >
  max_value | x < min_value) }, distance = 1.5)
}
\arguments{
\item{data}{a data frame containing only numerical features}

\item{func}{the function to evaluate for each point (must be boolean)}
}
\description{
Return a boolean matrix: evaluate the function func for each points of data
}
