% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric.mean.R
\name{geometric.mean}
\alias{geometric.mean}
\title{geometric.mean -- STOLEN FROM PSYCH PACKAGE.}
\usage{
geometric.mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{vector}

\item{na.rm}{flag to ignore NAs}
}
\seealso{
\code{\link{psych}}
}
\author{
William Revelle
}
