% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{greedy_model_selection}
\alias{greedy_model_selection}
\title{Compute the base detector greedy algorithm}
\usage{
greedy_model_selection(lor, t = 0.1)
}
\arguments{
\item{lor}{list of list of results named "probability_vector"}

\item{t}{threshold}
}
\description{
Compute the base detector greedy algorithm
}
