% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximity_based.R
\name{hclustering}
\alias{hclustering}
\title{Return an outlier score computed by hyerarchical clustering}
\usage{
hclustering(data, percent = 10, dist_meth = "euclidean")
}
\arguments{
\item{data}{a data frame containing only numerical features}

\item{percent}{number of clusters}
}
\description{
Return an outlier score computed by hyerarchical clustering
}
