% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQR.Broker.R
\name{IQR.Broker}
\alias{IQR.Broker}
\title{IQR Utility Broker}
\value{
A \code{list} of utility functions:
\itemize{
\item \code{IQR.from.quartiles(quartiles)}
\item \code{IQR.from.sample(sample)}
}
}
\description{
\code{IQR.Broker()} returns a set of utility functions that can be used to determine the inter quartile range from quartiles of a sample data set.
\itemize{
\item The first function: \code{IQR.from.quartiles}, calculates and returns the inter quartile range (IQR) from quartiles.
\item The second function: \code{IQR.from.sample}, calculates and returns the inter quartile range (IQR) from a sample data set.
}
}
