% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQR.Exceptions.R
\name{IQR.Exceptions}
\alias{IQR.Exceptions}
\title{IQR Exceptions}
\value{
A \code{list} of exceptions:
\itemize{
\item \code{argument.NULL()}
\item \code{type.mismatch()}
\item \code{property.missing()}
}
}
\description{
\code{IQR.Exceptions()} is an internal function that returns a set of exceptions used by \code{IQR.Validation()} service.
\itemize{
\item The first exception: \code{argument.NULL}, indicates to the user that no input argument was provided but required.
\item The second exception: \code{type.mismatch}, indicates to the user that the input is not of the correct data type.
\item The third exception: \code{property.missing}, indicates to the user that the input is missing a property.
}
}
