% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQR.Processor.R
\name{IQR.Processor}
\alias{IQR.Processor}
\title{IQR Processor}
\usage{
IQR.Processor(service)
}
\arguments{
\item{service}{IQR.Service instance injected into the processor.}
}
\value{
A \code{list} of processors:
\itemize{
\item \code{IQR()}
}
}
\description{
\code{IQR.Processor()} is an internal function that returns a processor used to calculate then \code{IQR}.
The processor includes dynamic dispatching to support different input types.
\itemize{
\item The first processor: \code{IQR.from.quartiles}, calculates the IQR from the first and third quartiles.
\item The second processor: \code{IQR.from.sample}, calculates the IQR from the sample data.
}
}
