% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQR.Validator.R
\name{IQR.Validator}
\alias{IQR.Validator}
\title{IQR Validator}
\value{
A \code{list} of validators:
\itemize{
\item \code{sample()}
\item \code{quartiles()}
\item \code{exist(argument)}
\item \code{is.numeric()}
\item \code{is.list()}
\item \code{has.first.quartile()}
\item \code{has.third.quartile()}
}
}
\description{
\code{IQR.Validator()} is an internal function that returns a set of validators used by \code{IQR.Service()}.
\itemize{
\item The first validator: \code{sample}, checks if the sample data provided is not NULL and is numeric.
\item The second validator: \code{quartiles}, checks if the quartiles is not NULL, a list and has first and third quartiles.
\item The third validator: \code{exist}, checks if a non-null data argument was provided.
\item The fourth validator: \code{is.numeric}, checks if the data argument provided contains numerical data.
\item The fifth validator: \code{is.list}, checks if the data argument provided is a list.
\item The sixth validator: \code{has.first.quartile}, checks if the quartiles list has a first quartile property.
\item The seventh validator: \code{has.third.quartile}, checks if the quartiles list has a third quartile property.
}
}
