% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximity_based.R
\name{LOCI}
\alias{LOCI}
\title{Return an outlier score based on the Local Correlation Integral}
\usage{
LOCI(data, alpha = c(0.2, 0.5, 0.8), vector = seq(from = 0.2, to = 0.8, by =
  0.3), only_mean = TRUE)
}
\arguments{
\item{data}{a data frame containing only numerical features}

\item{vector}{vector containing proportion of point in the neighboorhood of the point studied}
}
\description{
Return an outlier score based on the Local Correlation Integral
}
