% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximity_based.R
\name{MKP}
\alias{MKP}
\title{Compute a modified K-means}
\usage{
MKP(data, n, max_clusters, dist_meth = "euclidean")
}
\arguments{
\item{data}{a data frame containing only numerical features}

\item{n}{proportion of dataset use as cluster 0<n<1}

\item{max_clusters}{maximal proportion allowed
Initialize n clusters at random, or take n points in the dataset}
}
\description{
Compute a modified K-means
}
