% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximity_based.R
\name{MST2}
\alias{MST2}
\title{Minimum spanning tree}
\usage{
MST2(data, centroids, percent = 0.05, dist_meth = "euclidean")
}
\arguments{
\item{data}{a data frame containing only numerical features}

\item{centroids}{a matrix containing centroids}

\item{percent}{maximalsize of a group to consider all this points as outliers}
}
\description{
Minimum spanning tree
}
