% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiOut.R
\name{multiOut}
\alias{multiOut}
\title{multivariate outliers (via MCD)}
\usage{
multiOut(dat, rmdo_alpha = 0.8, rmdo_nsamp = "best", nsims = 100,
  nsamps = 250, alpha = 0.01)
}
\arguments{
\item{dat}{the data frame from which to identify outliers. Observation identifiers should be in the rownames. Assumes complete data and variables have been centred and standardized, if necessary.}

\item{rmdo_alpha}{the alpha to be used in CovMcd – the proportion of observations to keep in the robust cluster.}

\item{nsims}{number of repetitions of simulating multivariate normal in idOutliers.}

\item{nsamps}{number of observations to simulate within each repetition for idOutliers.}

\item{alpha}{the error to allow for quantile cut-offs. 1-alpha -> thisquant in idOutliers.}

\item{rmdo_samp}{the nsamp to be used in CovMcd – the maximum number of iterations to use for convergence of the robust cluster. "best" is recommended as "exhaustive enumeration is done, as long as the number of trials does not exceed 5000." See also ?\code{CovMcd}}
}
\value{
A list with two elements: outlier_decision and hidden_detail.
\item{outlier_decision}{ a data.frame indicating the observations that are outliers: 1 = outlier; 0 = not an outlier.}
hidden_detail is another list with the additional information:
\item{initDat}{the initial data supplied to the function.}
\item{RMDO}{data.frame of the outlier results in decreasing order of most-least outlying.Outlier=1 implies the observation is an outlier, while Outlier=0 implies otherwise.}
\item{RMDOalpha}{the alpha used in CovMcd, as supplied to the function by the user.}
\item{Contribs}{data.frame of the contribution for each observation-variable pair.}
\item{Cutoff}{the cut-off based on averaging the quantiles of each repetition.}
}
\description{
Calls getRMDO() and idOutliers() to calculate the robust parameters, RMDO, and identify the cutoff for outliers.  Then, compares the RMDO for each observation with the outlier cutoff.  It also reformats the contribution matrix.
}
\examples{
data(beer.tasting.notes)
the.data <- expo.scale(beer.tasting.notes$data)
the.data_corrected_for_ABV <- apply(the.data,2, function(x){ resid(lm(x~beer.tasting.notes$sup.data[,"ABV"])) } )
beer_mcd <- multiOut(the.data)
beer.corrected_mcd <- multiOut(the.data_corrected_for_ABV)
}

