% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.R
\name{one_class_svm}
\alias{one_class_svm}
\title{compute  an outlier score based on the distance from the support vector of a one-class SVM}
\usage{
one_class_svm(x, svm.nu = 0.1)
}
\arguments{
\item{data}{a data frame containing only numerical features}
}
\description{
compute  an outlier score based on the distance from the support vector of a one-class SVM
}
