% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outlier.Orchestrator.R
\name{Orchestrator}
\alias{Orchestrator}
\title{Outlier Orchestrator}
\usage{
Orchestrator()
}
\value{
A list containing two functions: \code{extract} and \code{remove}.
\itemize{
\item \code{extract}: Extracts outliers from the sample.
\item \code{remove}: Removes outliers from the sample.
}
}
\description{
\code{Outlier.Orchestrator()} is a function that orchestrates the outlier detection process.
}
