% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outlier.Broker.R
\name{Outlier.Broker}
\alias{Outlier.Broker}
\title{Outlier Broker}
\usage{
Outlier.Broker(boundary)
}
\arguments{
\item{boundary}{A Boundary.Service object that defines the upper and lower boundaries for outlier detection.}
}
\value{
A list containing two functions: \code{extract} and \code{remove}.
\itemize{
\item \code{extract}: Extracts outliers from the sample.
\item \code{remove}: Removes outliers from the sample.
}
}
\description{
This service provides methods to extract and remove outliers from a sample based on specified boundaries.
}
