% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outlier.Validator.R
\name{Outlier.Validator}
\alias{Outlier.Validator}
\title{Outlier Validator}
\value{
A \code{list} of validators:
\itemize{
\item \code{sample()}
\item \code{exist()}
\item \code{is.numeric()}
\item \code{is.medcouple.in.range()}
}
}
\description{
\code{Outlier.Validator()} is an internal function that returns a set of validators used by \code{Outlier.Validation()} service.
\itemize{
\item The first validator: \code{sample}, validates the sample data provided for analysis.
\item The second validator: \code{exist}, checks if the sample data is not NULL.
\item The third validator: \code{is.numeric}, checks if the sample data is numeric.
\item The fourth validator: \code{is.medcouple.in.range}, checks if the medcouple of the sample data is in the range of -0.6 to 0.6.
}
}
