% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outlier.Warnings.R
\name{Outlier.Warnings}
\alias{Outlier.Warnings}
\title{Outlier Warnings}
\value{
A \code{list} of warnings:
\itemize{
\item \code{out.of.range()}
}
}
\description{
\code{Outlier.Warnings()} is an internal function that returns a set of warnings used by \code{Outlier.Validation()} service.
\itemize{
\item The first warning: \code{out.of.range}, indicates to the user that the medcouple value is out of range.
}
}
