\name{outlieRs-package}
\alias{outlieRs-package}
\alias{outlieRs}
\encoding{UTF-8}
\docType{package}
\title{
outlieRs
}
\description{
outlieRs description
}
\details{
\tabular{ll}{
Package: \tab outlieRs\cr
Type: \tab Package\cr
Version: \tab 2.8.20\cr
Date: \tab 2013-12-09\cr
Depends: \tab R (>=3.3.0)
License: \tab GPL-2\cr
URL: \tab https://github.com/derekbeaton/outlieRs/commits/master\cr
}
}
\author{

}
\references{

}
\keyword{ package }
\keyword{ multivariate }
\seealso{

}
\examples{
#For more examples, see each individual function (as noted above).
}
