% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximity_based.R
\name{pca_lof}
\alias{pca_lof}
\title{Return an outlier score computed by doing a pca + lof}
\usage{
pca_lof(data, pca_meth = basic_pca)
}
\arguments{
\item{data}{a data frame containing only numerical features}
}
\description{
Return an outlier score computed by doing a pca + lof
}
