% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.R
\name{pca_score}
\alias{pca_score}
\title{compute  an outlier score based on the distance from a pca 
PCA from http://shahramabyari.com/2015/12/20/data-preparation-for-predictive-modeling-centering-scaling/}
\usage{
pca_score(x, pca = basic_pca)
}
\arguments{
\item{data}{a data frame containing only numerical features}
}
\description{
compute  an outlier score based on the distance from a pca 
PCA from http://shahramabyari.com/2015/12/20/data-preparation-for-predictive-modeling-centering-scaling/
}
