% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{preprocessing}
\alias{preprocessing}
\title{Preprocess the data, transform class into boolean, remove low variances, and apply asinh}
\usage{
preprocessing(data, preprocess_method = "boxcox", pca = FALSE)
}
\arguments{
\item{data}{dataset}
}
\description{
Preprocess the data, transform class into boolean, remove low variances, and apply asinh
}
