% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quartile.Broker.R
\name{Quartile.Broker}
\alias{Quartile.Broker}
\title{Quartile Utility Services}
\value{
A \code{list} of utility functions:
\itemize{
\item \code{first(data)}
\item \code{third(data)}
}
}
\description{
\code{Quartile.Service()} returns a set of utility functions that can be used to analyze data:
\itemize{
\item The first function: \code{first}, calculates and returns the 25th percentile of a sample data set.
\item The second function: \code{third}, calculates and returns the 75th percentile of a sample data set.
}
}
