% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outlier.Exposure.R
\name{remove.outliers}
\alias{remove.outliers}
\title{Remove outliers}
\usage{
remove.outliers(input, name)
}
\arguments{
\item{input}{A vector or data.frame from which to remove outliers.}

\item{name}{The name of the column in the data.frame which should be used for outlier removal. If \code{input} is a vector, this parameter is ignored.}
}
\value{
A vector or data.frame with no outliers.
}
\description{
\code{Outliers.remove()} is a function that removes outliers from a sample.
}
