% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scaling.Service.R
\name{Scaling.Service}
\alias{Scaling.Service}
\title{Scaling Utility Services}
\value{
A \code{list} of utility functions:
\itemize{
\item \code{upper(medcouple)}
\item \code{lower(medcouple)}
}
}
\description{
\code{Scaling.Service()} returns a pair of utility functions that is used to calculate a scaling factor as part of outlier detection:
\itemize{
\item The first function: \code{upper}, calculates and returns the upper scaling factor based on the medcouple value.
\item The second function: \code{lower}, calculates and returns the lower scaling factor based on the medcouple value.
}
}
\examples{
scale <- Scaling.Service()

upper.scaling <-
 1 |> scale[['upper']]()

rm(upper.scaling)
}
