% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scaling.Validator.R
\name{Scaling.Validator}
\alias{Scaling.Validator}
\title{Scaling Validator}
\value{
A \code{list} of validators:
\itemize{
\item \code{medcouple()}
\item \code{exist()}
\item \code{is.numeric()}
}
}
\description{
\code{Scaling.Validator()} is an internal function that returns a set of validators used by \code{Scaling.Service()}.
\itemize{
\item The first validator: \code{medcouple}, checks if the medcouple provided is not NULL and is numeric.
\item The second validator: \code{exist}, checks if a non-null medcouple argument was provided.
\item The third validator: \code{is.numeric}, checks if the medcouple argument provided contains numerical data.
}
}
