% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Skewness.Service.R
\name{Skewness.Service}
\alias{Skewness.Service}
\title{Skewness Utility Services}
\value{
A \code{list} of utility functions:
\itemize{
\item \code{medcouple(data)}
}
}
\description{
\code{Skewness.Service()} returns a set of utility functions that can be used to determine the skewness present in the sample data.
\itemize{
\item The first function: \code{medcouple}, calculates and returns the medcouple of a sample data set.
}
}
