% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{timeout}
\alias{timeout}
\title{add a timeout}
\usage{
timeout(fun, args, timeout = 60 * 60)
}
\arguments{
\item{fun}{the function to call}

\item{args}{the function arguments}
}
\description{
add a timeout
}
