% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_data_frame}
\alias{to_data_frame}
\title{a function to convert dataset to dataframes and to add row/column names if needed}
\usage{
to_data_frame(d, force = FALSE)
}
\arguments{
\item{data}{the original dataset}
}
\description{
a function to convert dataset to dataframes and to add row/column names if needed
}
