% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximity_based.R
\name{two_phases_clustering}
\alias{two_phases_clustering}
\title{Return an outlier score based on the two phases clustering algorithm}
\usage{
two_phases_clustering(data, n = 50, base_p_cluster = 0.1,
  max_p_clusters = 0.2)
}
\arguments{
\item{data}{a data frame containing only numerical features}

\item{n}{number of repetion of the algorithm}

\item{base_p_cluster}{initial number of clusters (proportion of data)}

\item{max_p_cluster}{maximum number of clusters (proportion of data)}
}
\description{
Return an outlier score based on the two phases clustering algorithm
}
