% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{variance}
\alias{variance}
\title{A function to compute the variance of the given feature data}
\usage{
variance(feature_data)
}
\arguments{
\item{feature_data}{a vector of values for a feature}
}
\description{
A function to compute the variance of the given feature data
}
\keyword{variance}
