% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outlier.Exposure.R
\name{extract.outliers}
\alias{extract.outliers}
\title{Extract outliers}
\usage{
extract.outliers(sample)
}
\arguments{
\item{sample}{A sample from which to extract outliers.}
}
\value{
A sample with outliers extracted.
}
\description{
\code{Outliers.extract()} is a function that extracts outliers from a sample.
}
