\name{outlieRs-package}
\alias{outlieRs-package}
\alias{outlieRs}
\encoding{UTF-8}
\docType{package}
\title{
outlieRs
}
\description{
outlieRs description
}
\details{
Package:  outlieRs\cr
Type:  Package\cr
Version: 2.8.20\cr
Date: 2013-12-09\cr
Depends:  R (>=3.3.0)
License: GPL-2\cr
URL: https://github.com/derekbeaton/outlieRs/commits/master\cr
}
\author{

}
\references{

}
\keyword{ package }
\keyword{ multivariate }
\seealso{

}
\examples{
#For more examples, see each individual function (as noted above).
}
