% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outlier.Exposure.R
\name{remove.outliers}
\alias{remove.outliers}
\title{Remove outliers}
\usage{
remove.outliers(sample)
}
\arguments{
\item{sample}{A sample from which to remove outliers.}
}
\value{
A sample with outliers removed.
}
\description{
\code{Outliers.remove()} is a function that removes outliers from a sample.
}
